% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-colnames.R
\name{check_colnames}
\alias{check_colnames}
\title{Check that colnames exist}
\usage{
check_colnames(col.names, time.name, qty.name = NULL, verbose = FALSE)
}
\arguments{
\item{time.name}{character vector of length one Name of the variable
containing time stamps for the observations.}

\item{qty.name}{character vector Name(s) of variable(s) in \code{data}
containing values observed quantities. If \code{qty.name = NULL}, the
default, all columns are retained.}

\item{verbose}{logical Report data columns found. Useful for debugging.}
}
\description{
Check that names passed as arguments for columns containing time stamps and
observed quantities exist in \code{col.names}.
}
\details{
If the column named according to \code{time.name} is not in
  \code{col.names} an error is issued. If \code{qty.name = NULL} all
  columns in \code{col.names} minus that in \code{time.name}. If
  \code{qty.name} is a \code{character} vector only names also present
  in \code{col.names} are returned, while if none is present, an error
  is issued. If only some names in \code{qty.name} are discarded a
  message is issued.
}
\keyword{internal}
