% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-summary.R
\name{chunk_summary}
\alias{chunk_summary}
\alias{tb_summary}
\title{Chunk summaries}
\usage{
chunk_summary(l, add.times = FALSE, tz = "UTC")

tb_summary(x)
}
\arguments{
\item{add.times}{logical If \code{TRUE}, list names are converted into
\code{POSIX.ct} values and added in a column named \code{time}.}

\item{tz}{character The time zone used to decode times.}

\item{x}{numeric vector.}
}
\value{
A numeric vector of length 9.
}
\description{
Compute statistical summaries for each numeric variable in a time series
chunk or in a list of time series chunks.
}
\keyword{internal}
