% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-integ.R
\name{combine_integ}
\alias{combine_integ}
\title{Combine time series differing in time step}
\usage{
combine_integ(time, var, new.integ.ms = 20)
}
\arguments{
\item{time}{numeric Vector of times from the time-series (\emph{x}-axis).}

\item{var}{numeric Vector of observations from the time-series (y-axis).}

\item{new.integ.ms}{numeric New integration time for the returned
time-series. Use same units as main time-series.}
}
\value{
A data frame
}
\description{
Combine time-series from higher resolution to lower resolution.
}
\details{
Timeseries of observations at different time steps, e.g. 10 ms and
30 ms are combined into a new data frame. Higher resolution needs to be a
multiple of lower resolution.
}
