% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise-chunk.R
\name{denoise_chunk}
\alias{denoise_chunk}
\title{Denoise differences within data frames}
\usage{
denoise_chunk(
  data,
  time.name = "TIMESTAMP",
  qty.name = NULL,
  ignore.range = 0.01,
  add.signs = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{data.frame or a list of data.frame objects Each data frame
containing at least one column with time stamps, one column with a measured
quantity, and one column of running differences for each measured quantity.}

\item{time.name}{character vector of length one Name of the variable
containing time stamps for the observations.}

\item{qty.name}{character vector Name(s) of variable(s) in \code{data}
containing values observed quantities. If \code{qty.name = NULL}, the
default, all columns are retained.}

\item{ignore.range}{numeric The multiplier to apply to the spread of \code{x}
to obtain the size of differences to ignore.}

\item{add.signs}{logical Flag indicating if values returned by
\code{\link{sign}()} on the de-noised differences are to be added to the
returned data frame chunks.}
}
\description{
Replace running differences smaller than a threshold by zeros in selected
columns of data frames.
}
\details{
When searching for changes in the sign of differences we may need to discard the
small values. These functions replace differences smaller than a threshold
by zeros. This is an alternative to smoothing that can be difficult to
implement for irregular time series.

The argument passed to \code{data} can be either a bare \code{data.frame}
object or a \code{list} containing one or more data frames, such as that
returned by \code{\link{split_chunks}()}.

Each column named in \code{qty.name}, if present, is replaced by the value
returned by function \code{\link{denoise_diffs}()}.
}
\seealso{
\code{\link{denoise_diffs}()}, \code{\link{split_cunks}()} and
  \code{\link{check_colnames}()}.
}
