% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-flecks.R
\name{find_flecks}
\alias{find_flecks}
\title{Find flecks}
\usage{
find_flecks(
  time,
  var,
  zero.lim = 5e-04,
  minTime = 0,
  minAmp = 0,
  minPdiff = 0,
  asymmetry = 1/4,
  trimCV = 0.05,
  asmMethod = c("mean", "max", "rm"),
  bounds = c(0, 1),
  timeSplit = 10,
  shadeflecks = FALSE,
  time.digits = 3,
  var.digits = 2,
  verbose = TRUE
)
}
\arguments{
\item{time}{numeric Vector of times from the time-series (\emph{x}-axis).}

\item{var}{numeric Vector of observations from the time-series (y-axis).}

\item{zero.lim}{numeric Limit for multiplication, only values higher than
\code{zero.lim} are kept (removes noise and prevents recording zeroes when
time-series is flat.).}

\item{minTime}{numeric Flecks found with duration below \code{minTime} will
be discarded. Keep at 0 to keep all flecks.}

\item{minAmp}{numeric Flecks found with amplitude below \code{minAmp} will be
discarded. Keep at 0 to keep all flecks.}

\item{minPdiff}{numeric Flecks found with a percent difference between peak
and baseline that is below \code{minPdiff} will be discarded. Keep at 0 to
keep all flecks.}

\item{asymmetry}{numeric Threshold value to qualify fleck as asymeric.
Asymetry happens when the baseline have large difference in their value.}

\item{trimCV}{Control trimming threshold. Fleck baselines are trimmed
iteratively based on the coefficient of variation between two points  at
each baseline side.}

\item{asmMethod}{character One of \code{"mean"}, \code{"max"}, \code{"rm"}.
Decides what to do with asymetric flecks \code{"mean"} averages the two
baselines, \code{"max"} keeps the largest baseline, \code{"rm"} discard the
asymetric fleck.}

\item{bounds}{numeric vector of length 2 For relative amplitude calculations,
normalize between 0 and 1 by default.}

\item{timeSplit}{integer Increase time-series data frequency by linear
interpolation. A value of 10 usually guarantee accuracy.}

\item{shadeflecks}{logical If true, run the function in \emph{shadefleck}
mode instead of in the default \emph{sunfleck} mode, i.e., the function
will find troughs instead of peaks in the data.}

\item{time.digits, var.digits}{integer Argument passed to parameter
\code{digits} in internal calls to \code{\link{round}()} for \code{time}
and \code{var} values.}

\item{verbose}{logical If \code{TRUE}, provides more information while
running.}
}
\description{
Detects and characterizes "flecks" in a time series of irradiances.
}
\details{
A sunfleck is characterized by an increase followed by a decrease
in irradiance. As a first step zero crossings of the derivative are located
using the same code as in \code{\link{find_zeros}()}. In a second step,
flecks are searched and once found, are checked for asymmetry
between baselines. If found, tries to extend baseline a bit. If still
asymmetric, behaviour is as defined by \code{asmMethod}. Then fleck is
compared againts criteria given by \code{minTime}, \code{minAmp} and
\code{minPdiff}. If passed, the baselines are trimmed. Conditions are
checked once more, and trimming is reversed if conditions are not passed
any more. In the last step overlapping flecks are removed and the time
interval between successive flecks is computed. Finally, the fleck properties
are returned in a data frame.
}
\references{
Durand M, Matule B, Burgess AJ, Robson TM. 2021. Sunfleck properties from
  time series of fluctuating light. \emph{Agricultural and Forest Meteorology}
  \strong{308-309}, 108554. \doi{10.1016/j.agrformet.2021.108554}
}
