% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-zeros.R
\name{find_zeros}
\alias{find_zeros}
\title{Find all timepoints when time series crosses zero}
\usage{
find_zeros(time, var, lim = 5e-04, timeSplit = 10, return_n1n2 = FALSE)
}
\arguments{
\item{time}{numeric Vector of times from the time-series (\emph{x}-axis).}

\item{var}{numeric Vector of observations from the time-series (y-axis).}

\item{lim}{numeric Limit for multiplication, only values higher than lim are
kept (removes noise and prevents recording zeroes when time-series is flat.).}

\item{timeSplit}{numeric Increase time-series frequency (value at t = 0 are
copied from t = 1 to t = 9, etc.). 10 usually garantees accuracy.}

\item{return_n1n2}{logical  If \code{TRUE}, also return the value of the
multiplication.}
}
\value{
An \code{integer} vector or a data frame with two variables.
}
\description{
First function to use on discrete time-series. Returns a vector giving the
position of every zero crossing within the time-series.
}
\details{
For every timepoint,
the numerical derivative is calculated using \code{\link{calDev}()} and each sequential
timepoint of the numerical derivative is multiplied. When the results is
negative, the time-series crosses zero.
}
\references{
Durand M, Matule B, Burgess AJ, Robson TM. 2021. Sunfleck properties from
  time series of fluctuating light. *Agricultural and Forest Meteorology*
  **308-309**, 108554. \doi{10.1016/j.agrformet.2021.108554}
}
