% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-data.R
\docType{data}
\name{three_chunks.tb}
\alias{three_chunks.tb}
\title{PAR irradiance time series}
\format{
A \code{"data.frame"} object with 54700 rows.
}
\usage{
three_chunks.tb
}
\description{
Time series of photosynthetically active radiation (PAR) measured at 50 ms
time intervals in 15 min-long bursts every 30 min.
}
\details{
The \code{data.frame} named \code{three_chunks.tb} contains
  time stamps and PAR photon irradiances (PPFD).

  The variables in each member spectrum are as follows:
  \itemize{ \item
  \code{time} \item \code{Q_PAR} }
}
\examples{

head(three_chunks.tb)

}
\keyword{datasets}
