% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise-chunks.R
\name{denoise_diffs}
\alias{denoise_diffs}
\title{Denoise one vector of differences}
\usage{
denoise_diffs(x, x.diff, ignore.range = 0.01)
}
\arguments{
\item{x}{numeric vector Supplying the range of data before computing
differences, or the data themselves.}

\item{x.diff}{numeric vector The running differences to be de noised.}

\item{ignore.range}{numeric The multiplier to apply to the spread of \code{x}
to obtain the size of differences to ignore.}
}
\description{
Denoise one vector of differences
}
