% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-flecks.R
\name{find_flecks}
\alias{find_flecks}
\title{Find flecks}
\usage{
find_flecks(
  time,
  var,
  zeroes = find_zeros(time, var),
  minTime = 0,
  minAmp = 0,
  minPdiff = 0,
  asymmetry = 1/4,
  trimCV = 0.05,
  asmMethod = c("mean", "max", "rm"),
  bounds = c(0, 1),
  timeSplit = 10,
  shadeflecks = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{time}{numeric Vector of times from the time-series (\emph{x}-axis).}

\item{var}{numeric Vector of observations from the time-series (y-axis).}

\item{zeroes}{Vector of zeroes returned by \code{find_zeroes()} on the same
\code{time} and \code{var} arguments.}

\item{minTime}{numeric Flecks found with duration below `minTime` will be
discarded. Keep at 0 to keep all flecks.}

\item{minAmp}{numeric Flecks found with amplitude below `minAmp` will be
discarded. Keep at 0 to keep all flecks.}

\item{minPdiff}{numeric Flecks found with a percent difference between peak
and baseline that is below `minPdiff` will be discarded. Keep at 0 to keep
all flecks.}

\item{asymmetry}{numeric Threshold value to qualify fleck as asymeric.
Asymetry happens when the baseline have large difference in their value.}

\item{trimCV}{Control trimming threshold. Fleck baselines are trimmed
iteratively based on the coefficient of variation between two points  at
each baseline side.}

\item{asmMethod}{character One of \code{"mean"}, \code{"max"}, \code{"rm"}.
Decides what to do with asymetric flecks \code{"mean"} averages the two
baselines, \code{"max"} keeps the largest baseline, \code{"rm"} discard the
asymetric fleck.}

\item{bounds}{numeric vector of length 2 For relative amplitude calculations,
normalize between 0 and 1 by default.}

\item{timeSplit}{integer Increase time-series data frequency by linear
interpolation. A value of 10 usually guarantee accuracy.}

\item{shadeflecks}{logical If true, run the function in \emph{shadefleck}
mode instead of in the default \emph{sunfleck} mode,
i.e., the function will find troughs instead of peaks in the data.}

\item{verbose}{logical If \code{TRUE}, provides more information while
running.}
}
\description{
Detects flecks from time series and zero vector returned by
\code{\link{find_zeros}(()}.
}
\details{
A sunfleck is caracterized by an increased followed by a decrease.
Function search zero vector for such events. Once found, checks for asymetry
between baselines. If found, tries to extend baseline a bit. If still
asymeric, behaviour is as defined by \code{asmMethod}. Then fleck is
checks for criteria as defined by \code{minTime}, \code{minAmp} and
\code{minPdiff}. If passed, the baselines are trimmed. Conditions are
checked once more, and trimming is reversed if conditions are not passed
anymore. Finally, log the fleck in a table returned at the end. At the end,
remove flecks overlapping, and calculate interval in time between two flecks.
}
\references{
Durand M, Matule B, Burgess AJ, Robson TM. 2021. Sunfleck properties from
  time series of fluctuating light. \emph{Agricultural and Forest Meteorology}
  \strong{308-309}, 108554. \doi{10.1016/j.agrformet.2021.108554}
}
