% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobiologyFlecks.R
\docType{package}
\name{photobiologyFlecks-package}
\alias{photobiologyFlecks}
\alias{photobiologyFlecks-package}
\title{photobiologyFlecks: Analysis of sun- and shade flecks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for the analysis of time series of natural light irradiance measured at high frequency. Denoising, and quantification of duration and amplitude of sun-, shade-, cloud- and wind flecks.
}
\references{
Durand M, Matule B, Burgess AJ, Robson TM. 2021. Sunfleck properties from
  time series of fluctuating light. \emph{Agricultural and Forest Meteorology}
  \strong{308-309}, 108554. \doi{10.1016/j.agrformet.2021.108554}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/photobiologyFlecks/}
  \item Report bugs at \url{https://github.com/aphalo/photobiologyFlecks/issues/}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

Authors:
\itemize{
  \item Maxime Durand \email{maxime.durand@universite-paris-saclay.fr} (\href{https://orcid.org/0000-0002-8991-3601}{ORCID})
}

}
