% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ts-fleck.R
\name{plot_ts_fleck}
\alias{plot_ts_fleck}
\title{Plot time series of all found flecks.}
\usage{
plot_ts_fleck(
  time,
  var,
  zeroes,
  fleck.data = FALSE,
  timeSplit = 10,
  ask = FALSE
)
}
\arguments{
\item{time}{numeric Vector of times from the time-series (\emph{x}-axis).}

\item{var}{numeric Vector of observations from the time-series (y-axis).}

\item{zeroes}{Vector of zeroes returned by \code{find_zeroes()} on the same
\code{time} and \code{var} arguments.}

\item{fleck.data}{data.frame Data frame returned by \code{find_flecks()} for
the same arguments as passed to \code{time}, \code{var} and \code{zeros}.}

\item{timeSplit}{integer Increase time-series data frequency by linear
interpolation. A value of 10 usually guarantee accuracy.}
}
\description{
Plot time series of all found flecks.
}
\details{
These functions plot the results of the fleck detection algorithm
using functions from 'graphics' package. Function \code{plot_ts_fleck_ez()}
is simpler than  \code{plot_ts_fleck()}.
}
\references{
Durand M, Matule B, Burgess AJ, Robson TM. 2021. Sunfleck properties from
  time series of fluctuating light. \emph{Agricultural and Forest Meteorology}
  \strong{308-309}, 108554. \doi{10.1016/j.agrformet.2021.108554}
}
