% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chunks.R
\name{split_chunks}
\alias{split_chunks}
\title{Split data frame into chunks}
\usage{
split_chunks(
  data,
  time.name = "TIMESTAMP",
  qty.name = NULL,
  step.len,
  chunk.min.rows = 2,
  add.diffs = TRUE,
  verbose = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{data.frame Containing at least one coloumn with time stamps and
one column with a measured quantity.}

\item{time.name}{character vector of length one Name of the variable
containing time stamps for the observations.}

\item{qty.name}{character vector Name(s) of variable(s) in \code{data}
containing values observed quantities. If \code{qty.name = NULL}, the
default, all columns are retained.}

\item{step.len}{numeric or duration Length of minimum time step length
between data chunks. If numeric, expressed in seconds.}

\item{chunk.min.rows}{integer The minimum number of rows that a chunk must
have not to be discarded.}

\item{add.diffs}{logical Flag indicating if values returned by
\code{\link{diff}()} are to be added to the returned data frame chunks.}

\item{verbose}{logical Report progress by printing times at gaps and number
of rows in each chunk.}

\item{na.rm}{logical Omit rows of \code{data} containing \code{NA} values
after selecting variables.}
}
\value{
A list of data frames of varying length, depending on the number of
  chunks found, possibly of length zero. The members of the list are named
  based on the starting time of each chunk. The variables included in the
  member data frames are those named by \code{time.name} and \code{qty.name}
  and optionally, their running differences.
}
\description{
Split a time series stored in a data frame at breaks (long time steps),
  returning a list of data frames or data chunks.
}
\details{
When time series of data are acquired in bursts or chunks separated
  by longer time intervals it can be useful to extract the chunks into
  separate data frames before further analysis. This implementation does not
  assume the same duration for all chunks or the gaps, it searches for time
  intervals longer than a threshold duration and splits the data at these
  points. If the data contains no gaps, the whole data is returned as a
  single chunk.

  When a minimum length for the individuals chunks is set with an argument to
  \code{chunk.min.rows}, chunks with fewer rows are discarded silently,
  unless \code{verbose = TRUE}.

  With \code{add.diffs = TRUE} the running differences between values in the
  current row and the one above are added to the returned data frames. The
  value in the first row is \code{NA} for running differences, except for
  the time, in which case it is the time difference to the precceeding value
  in \code{data}.

  Method \code{\link{diff}()} must be available for the class of the variable
  named by the argument to \code{time.name}. The class of this column is in
  most cases numeric, date, or time. If \code{add.diffs = TRUE} this
  requirement also applies to the variable(s) named by the argument passed to
  \code{qty.name}.

  The number of chunks in the returned list of data frames and their lengths
  are reported in a \code{\link{message}()}.
}
